Test For Projectile
I added a test scene that contains a tower and projectile. For the projectile I added a spawn point so when space is pressed the projectile will shoot towards the castle. This is to test shooting of the projectile and I also added the Launch Script to the projectile to see how the projectile hitting the tower would react. I did not see a way to test that otherwise since there was no way to shoot the projectile. Was not able to test however due to some networking issues. Not sure what the problem there is.